+++ lib/models/location.dart
@@ -1,0 +1,5 @@
+enum Location {
+  marinaSquare,
+  ricochetSentosa,
+  ricochetLaguna,
+}