+++ lib/router/routes.g.dart
@@ -13,6 +13,16 @@ List<RouteBase> get $appRoutes => [
 RouteBase get $homeRoute => GoRouteData.$route(
       path: '/',
       factory: $HomeRouteExtension._fromState,
+      routes: [
+        GoRouteData.$route(
+          path: 'users',
+          factory: $UsersRouteExtension._fromState,
+        ),
+        GoRouteData.$route(
+          path: 'events',
+          factory: $EventsRouteExtension._fromState,
+        ),
+      ],
     );
 
 extension $HomeRouteExtension on HomeRoute {
@@ -20,6 +30,40 @@ extension $HomeRouteExtension on HomeRoute {
 
   String get location => GoRouteData.$location(
         '/',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}
+
+extension $UsersRouteExtension on UsersRoute {
+  static UsersRoute _fromState(GoRouterState state) => const UsersRoute();
+
+  String get location => GoRouteData.$location(
+        '/users',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}
+
+extension $EventsRouteExtension on EventsRoute {
+  static EventsRoute _fromState(GoRouterState state) => const EventsRoute();
+
+  String get location => GoRouteData.$location(
+        '/events',
       );
 
   void go(BuildContext context) => context.go(location);