+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 - Adds `go_router`'s router.
+- Adds dummy events and users screens.
 
 ## 0.0.0